react-template-helper: Add two tests
One test for required `component` argument to `{{> React}}`, now passes

Another test for sibling detection when rendering a React component within
a container element with a conditional additional element. This test
fails since after the first React render Blaze loses track of the marker
nodes it uses to place newly added elements. We may not be able to do
sibling detection in runtime, unfortunately. This test is commented out.