Fixes Issue #4 -  Leaking React Components
Patch for issue #4 where a React component doesn’t get un-mounted
before Blaze destroys it’s own view.

Creates a random id and attaches it to the blaze wrapper container with
a data attribute. This value is stored in the template instance so it
can be
accessed at `onDestroy`.

Sample repo can be used to verify proper use. Inspector shows an empty
Blaze container div after un-mounting and before exiting `onDestroyed`

https://github.com/AdamBrodzinski/react-blaze-helper-fix

`PACKAGE_DIRS=` was used to link this patch to repo above.