Allow this.state and this.props in getMeteorData
This commit updates the examples, but not the guide!

getMeteorData() now takes no arguments.  You can access
this.props and this.state directly from it.  We achieve this by
temporarily swapping out this.props and this.state while
getMeteorData() is called.

The reason we had the arguments is that getMeteorData is called
from two lifecycle hooks, componentWillMount and componentWillUpdate,
and in the latter, the new props and state that will be used to
render have not been assigned yet.  However, our approach is modeled
on the proposed Observe API for React, which does not have this
problem because it hooks into the update cycle after
componentWillUpdate, and after the new props and state are assigned,
but before render().  Unfortunately, there's no lifecycle hook there
for us to use, but we can simulate it with a little harmless hacking
of this.props and this.state.