Make some dependency versions exact and explicit
react-runtime@versionX should explicitly depend on
react-runtime-{dev,prod}@versionX

Remove some `versionsFrom`, because it seems slightly clearer not to
have it when it doesn't apply to any of the dependencies.  Note that
versionsFrom will never apply from one of these packages to another,
because it only applies from a package not in a release to a package 
in a release.