Make task text properly reactive (fix #37 fix #43)
Changing defaultValue in React does not change input value.  You must
use `value` (or set it manually in the DOM).  If you use `value`, the
input element becomes "controlled" and anything the user types must be
processed by an onChange handler and incorporated into what's rendered,
or the entered text will be reverted.

If we simply pipe the data from minimongo via props into the input's
value, then minimongo controls the input value and React does not allow
us to display a value in the DOM that hasn't been committed to
minimongo, and thus we can't throttle writes to minimongo.
So, in order to let the user type but throttle updates, we use some
state, which is only used while the input is focused.

A possible improvement would be to define a `<LiveTextInput>` component
that encapsulates the state and throttling.  Another, more Flux-like
approach would be to somehow do the throttling in the data layer.
A third solution is to do what we do for list titles, and only use an
<input> element when the user is editing.