Make react-runtime-dev load before -prod
We had it the other way around (prod and then dev) in order to make
the React DevTools work, but it turns out we need dev to run first
in order to throw a good error when es5-shim is missing.  The checks
are not present in the prod version, which just crashes.