Don't serve prod copy of React in dev mode
We used to serve both dev and prod React in dev mode because we
didn't have the prodOnly package feature flag, but now we do
(on devel).

This lets us stop worrying about two side effects of serving two
copies of React: one, the prod version running before the dev
version checks for the required polyfills, and two, the prod version
overwriting the dev version as the copy of the library that gets to
talk to the DevTools.