A script that publishes all changed packages and dependencies
(The script doesn't yet actually call publish and merge
branches correctly)

This script scans the packages in this repo to see which
have changed compared to the 'master' branch. Then it
asks you for new versions for those packages, and updates
the version in any package that depends on those packages.