Fix lost reactivity in some cases.
If a computation runs, but the user's reactiveFn is not invoked, the computation will lose track of the reactive data sources. To avoid that, in cases where we will not run the user's reactiveFn (a reactive update which happens after first render, but before useEffect) we must always dispose of computation and recreate when the component is committed (useEffect runs). Thanks to @menelike for finding this.