Stop publishing index.d.ts as an asset file
So long as the `typescript` package is used by this package, then all
TypeScript files (including definition files) will get included in the
resulting build without needing to specify it as an asset. Including it
twice (both as a source and an asset file) confuses `zodern:types` and
causes it to skip over the package without including either.

We also don't need the package-types.json file, as we were never
including it in the isopack.