fix: #443 useSyncEffect Breaks useFind Reactivity Under Suspense
In the client-side implementation of suspense/useFind, useFindClient is used. PR #433 introduced useSyncEffect.
Nevertheless, when used in Strict Mode and with Suspense, the useSyncEffect implementation results in non-trivial and hard-to-predict issues.
As a result, the Suspense-enabled useFindSuspenseClient will no longer reference useFindClient and will instead be a standalone implementation (derived from the last stable version of useFindClient).